requirements.txt
requests>=2.25.0
numpy>=1.21.0
flask==2.0.1
qrcode==7.3.1
pillow==9.5.0
gunicorn==20.1.0
werkzeug==2.0.3
